requirements.txt
# Django Core
Django==4.2,<5.0
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-filter==23.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Celery & Async Tasks
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.3

# Authentication & Security
PyJWT==2.8.0
python-jose==3.3.0
bcrypt==4.1.2

# CORS & Middleware
django-cors-headers==4.3.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1

# Monitoring & Logging
sentry-sdk==1.40.6
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.0
model-bakery==1.16.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# API Documentation
drf-spectacular==0.27.1

# Performance
gunicorn==21.2.0
uvicorn==0.25.0

# Development Tools
ipython==8.20.0
black==24.1.1
flake8==7.0.0

# Deployment
whitenoise==6.6.0

# Email
django-email-verification==0.2.2

# Optional: Additional Utilities
python-slugify==8.0.4
shortuuid==1.0.11
